third_party/blink/web_tests/platform/generic/fast/js/var-declarations-expected.txt

This page tests var declarations inside various block structures and control statements.

Before executing blocks containing var declarations: 
-----
PASS: 'v1' in window should be true and is.
PASS: 'v2' in window should be true and is.
PASS: 'v3' in window should be true and is.
PASS: 'v4' in window should be true and is.
PASS: 'v5' in window should be true and is.
PASS: 'v6' in window should be true and is.
PASS: 'v7' in window should be true and is.
PASS: 'v8' in window should be true and is.
PASS: 'v9' in window should be true and is.

After executing blocks containing var declarations: 
-----
PASS: 'v1' in window should be true and is.
PASS: 'v2' in window should be true and is.
PASS: 'v3' in window should be true and is.
PASS: 'v4' in window should be true and is.
PASS: 'v5' in window should be true and is.
PASS: 'v6' in window should be true and is.
PASS: 'v7' in window should be true and is.
PASS: 'v8' in window should be true and is.
PASS: 'v9' in window should be true and is.
